apps/training/pyproject.toml
[project]
name = "training"
version = "0.1.0"
description = "Machine learning model training application"
requires-python = ">=3.12"
dependencies = [
    "fairlearn>=0.12.0",
    "joblib>=1.5.3",
    "matplotlib>=3.10.8",
    "ml-core",
    "mlflow>=3.10.0",
    "numpy>=2.4.1",
    "pandas>=2.2.0",
    "pillow>=12.1.0",
    "pymupdf>=1.26.7",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "scikit-learn>=1.3.0",
    "shared",
    "polars>=1.37.1",
    "scipy>=1.17.0",
    "xgboost>=3.1.3",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.1",
    "psycopg2-binary>=2.9.11",
    "sentence-transformers>=5.2.2",
    "tqdm>=4.65.0", # For progress bars in GraphQL scraper
    "httpx>=0.27.0", # For async HTTP in GraphQL scraper
    "torch>=2.10.0",
    "shap>=0.51.0",
    "google-cloud-storage>=3.10.0",
]


[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true  # Only use this index for packages that explicitly ask for it

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }
torchaudio = { index = "pytorch-cpu" }
ml-core = { workspace = true }
shared = { workspace = true }


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
